AdminDist {wwmetrics}	R Documentation

AdminDist

Description

Data on Administrative distances based on the Worldwide Governance Indicators by the World Bank.

Usage

data("AdminDist")

Format

A tibble with 786 173 observations and 12 variables:

ID. Factor indicating the countries forming each country pair (ISO3 code) and the year.

year. Integer indicating the year.

CountryPair. Factor indicating the country pair (ISO3 code).

origin. Factor indicating the first country (ISO3 code) of the country pair.

destination. Factor indicating the second country (ISO3 code) of the country pair.

AdminDist. Numeric. Administrative distance as the difference in the average of the scores obtained on the six following Worldwide Governance Indicators.

diffr_in_Control_of_corruption. Numeric. Difference in the score the World Bank granted the origin country and the destination country for each year on the indicator Control of Corruption.

diffr_in_Government_effectiveness. Numeric. Difference in the score the World Bank granted the origin country and the destination country for each year on the indicator Government Effectiveness.

diffr_in_Political_stability_and_absence_of_violence_terrorism. Numeric. Difference in the score the World Bank granted the origin country and the destination country for each year on the indicator Political Stability and Absence of Violence and Terrorism.

diffr_in_Regulatory_quality. Numeric. Difference in the score the World Bank granted the origin country and the destination country for each year on the indicator Regulatory Quality.

diffr_in_Rule_of_law. Numeric. Difference in the score the World Bank granted the origin country and the destination country for each year on the indicator Rule of Law.

diffr_in_Voice_and_accountability. Numeric. Difference in the score the World Bank granted the origin country and the destination country for each year on the indicator Voice and Accountability.

Source

https://datacatalog.worldbank.org/dataset/worldwide-governance-indicators

References

Worldwide Governance Indicators (www.govindicators.org)

[Package wwmetrics version 0.1.0 Index]