EIA {wwmetrics}	R Documentation

Economic Integration Agreements (EIA)

Description

Data on Economic Integration Agreements

Usage

data("EIA")

Format

A tibble with 2 383 290 observations and 6 variables:

ID. Factor indicating the countries forming each country pair (ISO3 code) and the year.

year. Integer indicating the year.

CountryPair. Factor indicating the country pair (ISO3 code).

origin. Factor indicating the first country (ISO3 code) of the country pair.

destination. Factor indicating the second country (ISO3 code) of the country pair.

EconomicIntegrationAgreements. Factor. From the website: "While most other similar data bases use a binary (0-1) variable to index the absence or presence of an EIA, our data base in Excel format uses a multichotomous index (0-6), with the following interpretations (described in detail in the data base): 0 denotes no existing Economic Integration Agreement, 1 denotes a One-Way Preferential Trade Agreement, 2 denotes a Two-Way Preferential Trade Agreement, 3 denotes a Free Trade Agreement, 4 denotes a Customs Union, 5 denotes a Common Market, 6 denotes an Economic Union"

Source

http://kellogg.nd.edu/nsf-kellogg-institute-data-base-economic-integration-agreements

References

NSF-Kellogg Institute Data Base on Economic Integration Agreements

[Package wwmetrics version 0.1.0 Index]