EconDist {wwmetrics}	R Documentation

Economic distance

Description

Data on Economic distance for 65484 country pairs (256 home countries and 256 host countries) from 1960 to 2016.

Usage

data("EconDist")

Format

A tibble with 2 428 512 observations and 6 variables:

ID. A character string indicating the country pair (ISO3 code) and the year.

year. Integer indicating the year.

CountryPair. Factor uniting the origin and the destination using "-".

origin. Factor indicating the first country (ISO3 code) of the country pair.

destination. Factor indicating the second country (ISO3 code) of the country pair.

Economic_distance. Numeric. Economic distance calculated between the two countries forming each country pair based on the differences in GDP in USD.

Source

https://data.worldbank.org/indicator/NY.GDP.MKTP.CD

References

http://www.worldbank.org/

[Package wwmetrics version 0.1.0 Index]