GeoDist {wwmetrics}	R Documentation

Geographic distances

Description

Data about the geographic distances in km between 50176 country pairs.

Usage

data("GeoDist")

Format

A tibble with 50176 observations (224 countries) and 5 variables:

origin. Factor indicating the first country (ISO3 code) of the country pair.

destination. Factor indicating the second country (ISO3 code) of the country pair.

CountryPair. Factor uniting the origin and the destination using "-".

GeoDist. Numeric. Number of kilometers between the two countries forming each country pair.

GeoDist_Capitals. Numeric. Number of kilometers between the capitals of each country forming each country pair.

Source

http://www.cepii.fr/CEPII/en/welcome.asp

References

Mayer, T. & Zignago, S. (2011) Notes on CEPII’s distances measures : the GeoDist Database CEPII Working Paper 2011-25

[Package wwmetrics version 0.1.0 Index]