Hofstede {wwmetrics}	R Documentation

Hofstede's cultural dimensions scores

Description

Data from Hofstede's cultural dimensions scores in 214 countries.

Usage

data("Hofstede")

Format

A tibble with 224 observations and 7 variables:

Country. Factor indicating the country (ISO3 code).

PowerDistance. Integer. Defined as "This dimension expresses the degree to which the less powerful members of a society accept and expect that power is distributed unequally. The fundamental issue here is how a society handles inequalities among people. People in societies exhibiting a large degree of Power Distance accept a hierarchical order in which everybody has a place and which needs no further justification. In societies with low Power Distance, people strive to equalise the distribution of power and demand justification for inequalities of power." (source: https://www.hofstede-insights.com/models/national-culture/)

Individualism. Integer. Defined as "The high side of this dimension, called Individualism, can be defined as a preference for a loosely-knit social framework in which individuals are expected to take care of only themselves and their immediate families. Its opposite, Collectivism, represents a preference for a tightly-knit framework in society in which individuals can expect their relatives or members of a particular ingroup to look after them in exchange for unquestioning loyalty. A society’s position on this dimension is reflected in whether people’s self-image is defined in terms of “I” or “we.”" (source: https://www.hofstede-insights.com/models/national-culture/)

Masculinity. Integer. Defined as "The Masculinity side of this dimension represents a preference in society for achievement, heroism, assertiveness, and material rewards for success. Society at large is more competitive. Its opposite, Femininity, stands for a preference for cooperation, modesty, caring for the weak and quality of life. Society at large is more consensus-oriented. In the business context Masculinity versus Femininity is sometimes also related to as “tough versus tender” cultures." (source: https://www.hofstede-insights.com/models/national-culture/)

UncertaintyAvoidance. Integer. Defined as "The Uncertainty Avoidance dimension expresses the degree to which the members of a society feel uncomfortable with uncertainty and ambiguity. The fundamental issue here is how a society deals with the fact that the future can never be known: should we try to control the future or just let it happen? Countries exhibiting strong UAI maintain rigid codes of belief and behaviour, and are intolerant of unorthodox behaviour and ideas. Weak UAI societies maintain a more relaxed attitude in which practice counts more than principles." (source: https://www.hofstede-insights.com/models/national-culture/)

LongTermOrientation. Numeric. Defined as "Every society has to maintain some links with its own past while dealing with the challenges of the present and the future. Societies prioritize these two existential goals differently. Societies who score low on this dimension, for example, prefer to maintain time-honoured traditions and norms while viewing societal change with suspicion. Those with a culture which scores high, on the other hand, take a more pragmatic approach: they encourage thrift and efforts in modern education as a way to prepare for the future." (source: https://www.hofstede-insights.com/models/national-culture/)

Indulgence. Numeric. Defined as "Indulgence stands for a society that allows relatively free gratification of basic and natural human drives related to enjoying life and having fun. Restraint stands for a society that suppresses gratification of needs and regulates it by means of strict social norms." (source: https://www.hofstede-insights.com/models/national-culture/)

Source

http://geerthofstede.com/research-and-vsm/dimension-data-matrix/

References

Geert Hofstede, Culture's Consequences: International Differences in Work-Related Values. Beverly Hills CA: Sage Publications, 1980

[Package wwmetrics version 0.1.0 Index]