Major_Languages_by_Country {wwmetrics}	R Documentation

Major languages by country

Description

Data about the most widely spoken languages in 120 countries (data uploaded on July 14th, 2012 by Douglas Dow).

Usage

data("Major_Languages_by_Country")

Format

A tibble with 120 observations and 6 variables:

Country. Factor indicating the country (ISO3 code).

Major.Language.1. Factor indicating the most spoken language in this country.

Major.Language.2. Factor indicating the second most spoken language in this country (if applicable).

Major.Language.3. Factor indicating the third most spoken language in this country (if applicable).

Major.Language.4. Factor indicating the fourth most spoken language in this country (if applicable).

Major.Language.5. Factor indicating the fifth most spoken language in this country (if applicable).

Details

Defining Major Languages:

A major language for a given country is defined as any language which can be spoken by more than 20 For the countries used in our analyses, 113 languages qualified as a major language for at least one of the 120 countries. These languages have been grouped into a hierarchy of families, branches, 1st level sub-branches and 2nd level sub-branches based on Gordon's (2005) more substantial classification of 6,912 languages. Two Excel files documenting the hierarchy of language families and the major languages for each country are attached at the bottom of this page.

Source

https://sites.google.com/site/ddowresearch/home/scales/languages

References

Douglas Dow, https://sites.google.com/site/ddowresearch/home, Gordon, R. G. (ed), Ethnologue: Languages of the World, 2005

[Package wwmetrics version 0.1.0 Index]