Schwartz {wwmetrics}	R Documentation

The 7 Schwartz cultural value orientation scores for 75 countries

Description

Data from Schwartz's cultural value orientation scores in 75 countries.

Usage

data("Schwartz")

Format

A tibble with 80 observations and 9 variables:

Notes. Character string (only for some of the observations) to distinguish between different populations in a country (e.g., French-speaking Canadians and English-Speaking Canadians)

Country. Factor indicating the country (ISO3 code).

harmony. Numeric. "harmony emphasizes fitting into the world as it is, trying to understand and appreciate rather than to change, direct, or to exploit. Important values in harmony cultures include world at peace, unity with nature, and protecting the environment." (source: Schwartz, Shalom H. 2006b. "A Theory of Cultural Value Orientations: Explication and Applications." Comparative Sociology 5:137-82.)

embeddedness. Numeric. "In embeddedness cultures, people are viewed as entities embedded in the collectivity. Meaning in life is expected to come largely through social relationships, through identifying with the group, participating in its shared way of life, and striving toward its shared goals. Embedded cultures emphasize maintaining the status quo and restraining actions that might disrupt in-group solidarity or the traditional order." Available from: https://www.researchgate.net/publication/304715744_The_7_Schwartz_cultural_value_orientation_scores_for_80_countries [accessed Feb 16 2018].

hierarchy. Numeric. "Hierarchy cultures rely on hierarchical systems of ascribed roles to insure responsible, productive behavior. They define the unequal distribution of power, roles, and resources as legitimate and even desirable. People are socialized to take the hierarchical distribution of roles for granted, to comply with the obligations and rules attached to their roles, to show deference to superiors and expect deference from subordinates." Available from: https://www.researchgate.net/publication/304715744_The_7_Schwartz_cultural_value_orientation_scores_for_80_countries [accessed Feb 16 2018].

mastery. Numeric. "Mastery is the polar cultural response to [harmony]. It encourages active self-assertion in order to master, direct, and change the natural and social environment to attain group or personal goals. Values such as ambition, success, daring, and competence are especially important in mastery cultures." (source: Schwartz, Shalom H. 2006b. "A Theory of Cultural Value Orientations: Explication and Applications." Comparative Sociology 5:137-82.)

affective_autonomy. Numeric. "Affective autonomy encourages individuals to pursue affectively positive experience for themselves." Available from: https://www.researchgate.net/publication/304715744_The_7_Schwartz_cultural_value_orientation_scores_for_80_countries [accessed Feb 16 2018].

intellectual_autonomy. Numeric. "Intellectual autonomy encourages individuals to pursue their own ideas and intellectual directions independently." Available from: https://www.researchgate.net/publication/304715744_The_7_Schwartz_cultural_value_orientation_scores_for_80_countries [accessed Feb 16 2018].

egalitarianism. Numeric. "Egalitarian cultures seek to induce people to recognize one another as moral equals who share basic interests as human beings. They try to socialize their members to internalize a commitment to cooperate and to feel concern for everyone's welfare. People are expected to act for the benefit of others as a matter of choice." Available from: https://www.researchgate.net/publication/304715744_The_7_Schwartz_cultural_value_orientation_scores_for_80_countries [accessed Feb 16 2018].

Source

https://www.researchgate.net/publication/304715744_The_7_Schwartz_cultural_value_orientation_scores_for_80_countries

References

Schwartz, Shalom H. 2006. "A Theory of Cultural Value Orientations: Explication and Applications." Comparative Sociology 5 (2-3): 137-182

[Package wwmetrics version 0.1.0 Index]