TheGoodCountry {wwmetrics}	R Documentation

The Good Country Index

Description

Data from The Good Country Index: "The idea of the Good Country Index is simple: to measure what each country on earth contributes to the common good of humanity, and what it takes away, relative to its size. Using a wide range of data from the U.N. and other international organisations, we’ve given each country a balance-sheet to show at a glance whether it’s a net creditor to mankind, a burden on the planet, or something in between." (source: https://goodcountry.org/index/about-the-index)

Usage

data("TheGoodCountry")

Format

A tibble with 163 observations and 9 variables:

Country. Factor indicating the country (ISO3 code).

science_and_technology. Ordered factor indicating the rank of the country in terms of "Contributions to Science & Technology. 1. International students: Number of foreign students studying in the country (according to UNESCO) relative to the size of the economy. 2. Journal exports: Exports of periodicals, scientific journals and newspapers (according to ITC) relative to the size of the economy. 3. International publications: Number of articles published in international journals (according to SCImago) relative to the size of the economy. 4. Nobel prizes: Accumulated Nobel prizes (up to 2014) assigned to countries based on laureates' country of birth as well as country (countries) of institutional affiliation at the time of the award, relative to the size of the economy. 5. Patents: Number of International Patent Cooperation Treaty applications (according to WIPO) relative to the size of the economy." (more details [here]https://docs.google.com/spreadsheets/d/1o5DPYXoGmDEtqEZhGgvBDcXeEaVieMeSkFYoo9BiSXM/edit#gid=1206169975)

culture. Ordered factor indicating the rank of the country in terms of "Contributions to Culture. 6. Creative goods exports: Exports of creative goods (UNCTAD's Creative Economy Report categorisation) relative to the size of the economy. 7. Creative services exports: Exports of creative services (according to ITC's "trade in services" categories 10 and 11) relative to the size of the economy. 8. UNESCO dues in arrears as

international_peace_and_security. Ordered factor indicating the rank of the country in terms of "Contributions to International Peace and Security. 11. Peacekeeping troops: Number of peacekeeping troops sent overseas for UN missions, relative to the size of the economy. 12. Dues in arrears to UN peace keeping budgets as

world_order. Ordered factor indicating the rank of the country in terms of "Contributions to World Order. 16. Charity giving: Percentage of population that gives to charity (according to Charities Aid Foundation) as proxy for cosmopolitan attitude. 17. Refugees hosted: Number of refugees hosted (according to UNHCR) relative to the size of the economy. 18. Refugees generated: Number of refugees overseas (according to UNHCR) relative to the size of the population (negative indicator). 19. Birth rate: Population birth rate (according to World Bank as negative indicator). 20. UN Treaties signed: Number of UN treaties signed (up to 2014) as proxy for diplomatic action and peaceful conflict resolution." (more details [here]https://docs.google.com/spreadsheets/d/1o5DPYXoGmDEtqEZhGgvBDcXeEaVieMeSkFYoo9BiSXM/edit#gid=1206169975)(more details [here]https://docs.google.com/spreadsheets/d/1o5DPYXoGmDEtqEZhGgvBDcXeEaVieMeSkFYoo9BiSXM/edit#gid=1206169975)

planet_and_climate. Ordered factor indicating the rank of the country in terms of "Contributions to Planet and Climate. 21. Ecological footprint (2013): National Footprint Accounts, Ecological footprint (according to Global Footprint Network) relative to the size of the economy (negative indicator). 22. Reforestation since 1992: Re-/De-forestation since 1992 (according to UNEP Environmental Data Explorer). 23. Hazardous pesticides exports: Hazardous pesticides exports (according to UNEP Environmental Data Explorer) relative to the size of the economy (negative indicator). 24. CO2 emissions (2013): CO2 emissions (according to World Bank) relative to the size of the economy (negative indicator). 25. Ozone: Consumption of ozone-depleting substances - all (according to UNEP Environmental Data Explorer) relative to the size of the economy (negative indicator)." (more details [here]https://docs.google.com/spreadsheets/d/1o5DPYXoGmDEtqEZhGgvBDcXeEaVieMeSkFYoo9BiSXM/edit#gid=1206169975)

prosperity_and_equality. Ordered factor indicating the rank of the country in terms of "Contributions to Prosperity and Equality. 26. Open trading: Trading across borders (open trading performance compared to best practice; i.e. IFC distance to frontier). 27. UN volunteers abroad: Number of aid workers and volunteers sent overseas (according to UNV) relative to the size of the economy. 28. Fairtrade market size: Fairtrade market size (according to Fairtrade International) relative to the size of the economy. 29. FDI outflows: FDI outflow (according to UNCTAD) relative to the size of the economy. 30. Development assistance: Development cooperation contributions (aid according to Development Initiatives) relative to the size of the economy." (more details [here]https://docs.google.com/spreadsheets/d/1o5DPYXoGmDEtqEZhGgvBDcXeEaVieMeSkFYoo9BiSXM/edit#gid=1206169975)

health_and_wellbeing. Ordered factor indicating the rank of the country in terms of "Contributions to Health and Wellbeing. 31. Food aid: Food aid funding (according to WFP) relative to the size of the economy. 32. Pharmaceutical exports: Exports of pharmaceuticals (according to ITC) relative to the size of the economy. 33. Voluntary excess donations to the WHO: Voluntary excess contributions to World Health Organisation relative to the size of the economy. 34. Humanitarian aid donations: Humanitarian aid contributions (according to UNOCHA) relative to the size of the economy. 35. International Health Regulations Compliance: International Health Regulations Compliance (according to WHO)." (more details [here]https://docs.google.com/spreadsheets/d/1o5DPYXoGmDEtqEZhGgvBDcXeEaVieMeSkFYoo9BiSXM/edit#gid=1206169975)

overall_rank. Ordered factor indicating the rank of the country in terms of "The overall rank is based on the average of the category ranks. All data is 2014 data unless otherwise indicated. The actual overall rank is based on 163 countries included in the index and the mean rank score for the 7 category ranks." (more details [here]https://docs.google.com/spreadsheets/d/1o5DPYXoGmDEtqEZhGgvBDcXeEaVieMeSkFYoo9BiSXM/edit#gid=1206169975)

Source

https://goodcountry.org/index/about-the-index

References

[Simon Anholt]https://goodcountry.org/

[Package wwmetrics version 0.1.0 Index]