WGI {wwmetrics}	R Documentation

Worldwide Governance Indicators

Description

Data about the Worldwide Governance Indicators in 214 countries from 1996 to 2016.

Usage

data("WGI")

Format

A tibble with 3852 observations and 10 variables:

ID. A character string indicating the country (ISO3 code) and the year.

Country. Factor indicating the country (ISO3 code).

year. Integer indicating the year.

Control_of_corruption. Numeric. Score the World Bank granted the country for this year on the indicator control of corruption.

Government_effectiveness. Numeric. Score the World Bank granted the country for this year on the indicator government effectiveness.

Political_stability_and_absence_of_violence_terrorism. Numeric. Score the World Bank granted the country for this year on the indicator Political stability and absence of violence terrorism.

Rule_of_law. Numeric. Score the World Bank granted the country for this year on the indicator Rule of law.

Regulatory_quality. Numeric. Score the World Bank granted the country for this year on the indicator Regulatory quality.

Voice_and_accountability. Numeric. Score the World Bank granted the country for this year on the indicator Voice and accountability.

mean_WGI. Numeric. Average score for each country-year across all six previously mentioned indicators.

Source

https://datacatalog.worldbank.org/dataset/worldwide-governance-indicators

References

Worldwide Governance Indicators (www.govindicators.org)

[Package wwmetrics version 0.1.0 Index]